search/serializers.py
"""
Book: Building RESTful Python Web Services
Chapter 2: Working with class based views and hyperlinked APIs in Django
Author: Gaston C. Hillar - Twitter.com/gastonhillar
Publisher: Packt Publishing Ltd. - http://www.packtpub.com
"""
from rest_framework import serializers
from search.models import Reviews


class ReviewSerializer(serializers.ModelSerializer):
    class Meta:
        model = Reviews
        fields = ('productId', 
                  'userId', 
                  'profileName',
                  'helpfulness', 
                  'score',
                  'time',
                  'summary',
                  'text',
                  'search_score')
