requirements.txt
awsebcli==3.10.3
botocore==1.5.80
cement==2.8.2
colorama==0.3.7
Django==1.11.3
django-debug-toolbar==1.8
djangorestframework==3.6.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
jmespath==0.9.3
pathspec==0.5.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.2
PyYAML==3.12
RandomWords==0.2.1
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
sqlparse==0.2.3
tabulate==0.7.5
termcolor==1.1.0
websocket-client==0.44.0
